Add basic functionality testing of the API
In order to test the API, the mux has to parse the URI, therefore it
needs to be serve the "fake" requests after being initialized as normal.
Therefore it needs to be extracted into a dedicated struct.

Test a sequence of creating a list, viewing it, adding an item and
viewing it again. During each step, the http Status code is checked as
well as potential return messages.