Add ability to create new lists to website.
This involves a REST call from the browser to the API to create the list
and adding functionality to the web server to display that list.

The web server renders the list-view page, after performing a REST call
to obtain the data. This minimizes the amount of JS i have to write :)